fix the overflow in chacha

+++ iris-mpc-gpu/src/rng/chacha.cu
@@ -101,7 +101,7 @@ __device__ void chacha12_internal(uint32_t *buffer, uint32_t *d_state,
   int idx = blockIdx.x * blockDim.x + threadIdx.x;
   // the 64-bit counter part is in state[12] and 13, we add our local counter =
   // idx here may not overflow, caller has to ensure that
-  uint64_t counter = state[12] | (state[13] << 32);
+  uint64_t counter = state[12] | ((uint64_t)(state[13]) << 32);
   counter += idx;
   thread_state[12] = counter & 0xFFFFFFFF;
   thread_state[13] = counter >> 32;
